requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.1
pydantic==2.10.6
python-dotenv==1.0.0
pypdf==4.0.1
python-multipart==0.0.9
aiohttp==3.9.3 
azure-storage-blob==12.25.1
bcrypt==4.1.2
sendgrid==6.12.4
azure-search-documents==11.5.3
email_validator==2.2.0
sarvamai==0.1.17
vapi-python==0.1.0
vapi-server-sdk==1.7.2
azure-ai-documentintelligence==1.0.2
numpy==2.3.3
PyJWT==2.10.1
python-jose==3.5.0
pytz
sib-api-v3-sdk==7.6.0